Fix prmocode generator for carnet event type
